requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
neo4j==5.17.0
openai==1.12.0
sentence-transformers==2.3.1
pydantic==2.6.0
python-dotenv==1.0.0
httpx==0.26.0
